scripts/python/dj-scripts/spotify-to-mp3/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dacite==1.8.1
Deprecated==1.2.14
et_xmlfile==2.0.0
fastapi==0.103.2
h11==0.14.0
idna==3.7
jaconv==0.3.4
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.47.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
platformdirs==4.3.6
pycryptodomex==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pykakasi==2.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytube==15.0.0
pytz==2025.2
RapidFuzz==3.12.1
redis==5.0.7
requests==2.32.3
requests-futures==1.0.2
rich==13.9.4
setuptools==69.5.1
sherlock-project==0.15.0
six==1.16.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.5
spotdl==4.4.3
spotipy==2.24.0
starlette==0.27.0
stem==1.8.2
syncedlyrics==1.0.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.23.2
websockets==14.2
wrapt==1.16.0
yt-dlp==2025.10.14
ytmusicapi==1.11.1
